+++ lib/component/component.dart
@@ -48,7 +48,7 @@ class _PushStateState extends State<PushSwitchState> {
   void initState() {
     super.initState();
     addPostFrameCallback((_) {
-      push = BHP().getBool(UConstant.isPush) ?? true;
+      push = BHP().getBool(UConst.isPush) ?? true;
       setState(() {});
     });
   }
@@ -59,7 +59,7 @@ class _PushStateState extends State<PushSwitchState> {
       onChanged: (value) {
         if (value == push) return;
         push = value;
-        BHP().setBool(UConstant.isPush, push);
+        BHP().setBool(UConst.isPush, push);
         setState(() {});
       });
 }